+++ include/math.h
@@ -21,7 +21,7 @@ extern "C" {
 
 #define HUGE_VALF INFINITY
 #define HUGE_VAL  ((double)INFINITY)
-#define HUGE_VALL ((long double)INFINITY)
+#define HUGE_VALL ((double)INFINITY)
 
 #define MATH_ERRNO  1
 #define MATH_ERREXCEPT 2
@@ -38,7 +38,7 @@ extern "C" {
 
 int __fpclassify(double);
 int __fpclassifyf(float);
-int __fpclassifyl(long double);
+int __fpclassifyl(double);
 
 static __inline unsigned __FLOAT_BITS(float __f)
 {
@@ -80,7 +80,7 @@ static __inline unsigned long long __DOUBLE_BITS(double __f)
 
 int __signbit(double);
 int __signbitf(float);
-int __signbitl(long double);
+int __signbitl(double);
 
 #define signbit(x) ( \
 	sizeof(x) == sizeof(float) ? (int)(__FLOAT_BITS(x)>>31) : \
@@ -95,19 +95,19 @@ static __inline int __is##rel(type __x, type __y) \
 
 __ISREL_DEF(lessf, <, float_t)
 __ISREL_DEF(less, <, double_t)
-__ISREL_DEF(lessl, <, long double)
+__ISREL_DEF(lessl, <, double)
 __ISREL_DEF(lessequalf, <=, float_t)
 __ISREL_DEF(lessequal, <=, double_t)
-__ISREL_DEF(lessequall, <=, long double)
+__ISREL_DEF(lessequall, <=, double)
 __ISREL_DEF(lessgreaterf, !=, float_t)
 __ISREL_DEF(lessgreater, !=, double_t)
-__ISREL_DEF(lessgreaterl, !=, long double)
+__ISREL_DEF(lessgreaterl, !=, double)
 __ISREL_DEF(greaterf, >, float_t)
 __ISREL_DEF(greater, >, double_t)
-__ISREL_DEF(greaterl, >, long double)
+__ISREL_DEF(greaterl, >, double)
 __ISREL_DEF(greaterequalf, >=, float_t)
 __ISREL_DEF(greaterequal, >=, double_t)
-__ISREL_DEF(greaterequall, >=, long double)
+__ISREL_DEF(greaterequall, >=, double)
 
 #define __tg_pred_2(x, y, p) ( \
 	sizeof((x)+(y)) == sizeof(float) ? p##f(x, y) : \
@@ -122,231 +122,231 @@ __ISREL_DEF(greaterequall, >=, long double)
 
 double      acos(double);
 float       acosf(float);
-long double acosl(long double);
+double acosl(double);
 
 double      acosh(double);
 float       acoshf(float);
-long double acoshl(long double);
+double acoshl(double);
 
 double      asin(double);
 float       asinf(float);
-long double asinl(long double);
+double asinl(double);
 
 double      asinh(double);
 float       asinhf(float);
-long double asinhl(long double);
+double asinhl(double);
 
 double      atan(double);
 float       atanf(float);
-long double atanl(long double);
+double atanl(double);
 
 double      atan2(double, double);
 float       atan2f(float, float);
-long double atan2l(long double, long double);
+double atan2l(double, double);
 
 double      atanh(double);
 float       atanhf(float);
-long double atanhl(long double);
+double atanhl(double);
 
 double      cbrt(double);
 float       cbrtf(float);
-long double cbrtl(long double);
+double cbrtl(double);
 
 double      ceil(double);
 float       ceilf(float);
-long double ceill(long double);
+double ceill(double);
 
 double      copysign(double, double);
 float       copysignf(float, float);
-long double copysignl(long double, long double);
+double copysignl(double, double);
 
 double      cos(double);
 float       cosf(float);
-long double cosl(long double);
+double cosl(double);
 
 double      cosh(double);
 float       coshf(float);
-long double coshl(long double);
+double coshl(double);
 
 double      erf(double);
 float       erff(float);
-long double erfl(long double);
+double erfl(double);
 
 double      erfc(double);
 float       erfcf(float);
-long double erfcl(long double);
+double erfcl(double);
 
 double      exp(double);
 float       expf(float);
-long double expl(long double);
+double expl(double);
 
 double      exp2(double);
 float       exp2f(float);
-long double exp2l(long double);
+double exp2l(double);
 
 double      expm1(double);
 float       expm1f(float);
-long double expm1l(long double);
+double expm1l(double);
 
 double      fabs(double);
 float       fabsf(float);
-long double fabsl(long double);
+double fabsl(double);
 
 double      fdim(double, double);
 float       fdimf(float, float);
-long double fdiml(long double, long double);
+double fdiml(double, double);
 
 double      floor(double);
 float       floorf(float);
-long double floorl(long double);
+double floorl(double);
 
 double      fma(double, double, double);
 float       fmaf(float, float, float);
-long double fmal(long double, long double, long double);
+double fmal(double, double, double);
 
 double      fmax(double, double);
 float       fmaxf(float, float);
-long double fmaxl(long double, long double);
+double fmaxl(double, double);
 
 double      fmin(double, double);
 float       fminf(float, float);
-long double fminl(long double, long double);
+double fminl(double, double);
 
 double      fmod(double, double);
 float       fmodf(float, float);
-long double fmodl(long double, long double);
+double fmodl(double, double);
 
 double      frexp(double, int *);
 float       frexpf(float, int *);
-long double frexpl(long double, int *);
+double frexpl(double, int *);
 
 double      hypot(double, double);
 float       hypotf(float, float);
-long double hypotl(long double, long double);
+double hypotl(double, double);
 
 int         ilogb(double);
 int         ilogbf(float);
-int         ilogbl(long double);
+int         ilogbl(double);
 
 double      ldexp(double, int);
 float       ldexpf(float, int);
-long double ldexpl(long double, int);
+double ldexpl(double, int);
 
 double      lgamma(double);
 float       lgammaf(float);
-long double lgammal(long double);
+double lgammal(double);
 
 long long   llrint(double);
 long long   llrintf(float);
-long long   llrintl(long double);
+long long   llrintl(double);
 
 long long   llround(double);
 long long   llroundf(float);
-long long   llroundl(long double);
+long long   llroundl(double);
 
 double      log(double);
 float       logf(float);
-long double logl(long double);
+double logl(double);
 
 double      log10(double);
 float       log10f(float);
-long double log10l(long double);
+double log10l(double);
 
 double      log1p(double);
 float       log1pf(float);
-long double log1pl(long double);
+double log1pl(double);
 
 double      log2(double);
 float       log2f(float);
-long double log2l(long double);
+double log2l(double);
 
 double      logb(double);
 float       logbf(float);
-long double logbl(long double);
+double logbl(double);
 
 long        lrint(double);
 long        lrintf(float);
-long        lrintl(long double);
+long        lrintl(double);
 
 long        lround(double);
 long        lroundf(float);
-long        lroundl(long double);
+long        lroundl(double);
 
 double      modf(double, double *);
 float       modff(float, float *);
-long double modfl(long double, long double *);
+double modfl(double, double *);
 
 double      nan(const char *);
 float       nanf(const char *);
-long double nanl(const char *);
+double nanl(const char *);
 
 double      nearbyint(double);
 float       nearbyintf(float);
-long double nearbyintl(long double);
+double nearbyintl(double);
 
 double      nextafter(double, double);
 float       nextafterf(float, float);
-long double nextafterl(long double, long double);
+double nextafterl(double, double);
 
-double      nexttoward(double, long double);
-float       nexttowardf(float, long double);
-long double nexttowardl(long double, long double);
+double      nexttoward(double, double);
+float       nexttowardf(float, double);
+double nexttowardl(double, double);
 
 double      pow(double, double);
 float       powf(float, float);
-long double powl(long double, long double);
+double powl(double, double);
 
 double      remainder(double, double);
 float       remainderf(float, float);
-long double remainderl(long double, long double);
+double remainderl(double, double);
 
 double      remquo(double, double, int *);
 float       remquof(float, float, int *);
-long double remquol(long double, long double, int *);
+double remquol(double, double, int *);
 
 double      rint(double);
 float       rintf(float);
-long double rintl(long double);
+double rintl(double);
 
 double      round(double);
 float       roundf(float);
-long double roundl(long double);
+double roundl(double);
 
 double      scalbln(double, long);
 float       scalblnf(float, long);
-long double scalblnl(long double, long);
+double scalblnl(double, long);
 
 double      scalbn(double, int);
 float       scalbnf(float, int);
-long double scalbnl(long double, int);
+double scalbnl(double, int);
 
 double      sin(double);
 float       sinf(float);
-long double sinl(long double);
+double sinl(double);
 
 double      sinh(double);
 float       sinhf(float);
-long double sinhl(long double);
+double sinhl(double);
 
 double      sqrt(double);
 float       sqrtf(float);
-long double sqrtl(long double);
+double sqrtl(double);
 
 double      tan(double);
 float       tanf(float);
-long double tanl(long double);
+double tanl(double);
 
 double      tanh(double);
 float       tanhf(float);
-long double tanhl(long double);
+double tanhl(double);
 
 double      tgamma(double);
 float       tgammaf(float);
-long double tgammal(long double);
+double tgammal(double);
 
 double      trunc(double);
 float       truncf(float);
-long double truncl(long double);
+double truncl(double);
 
 
 #if defined(_XOPEN_SOURCE) || defined(_BSD_SOURCE)
@@ -408,19 +408,19 @@ float       ynf(int, float);
 #endif
 
 #ifdef _GNU_SOURCE
-long double lgammal_r(long double, int*);
+double lgammal_r(double, int*);
 
 void        sincos(double, double*, double*);
 void        sincosf(float, float*, float*);
-void        sincosl(long double, long double*, long double*);
+void        sincosl(double, double*, double*);
 
 double      exp10(double);
 float       exp10f(float);
-long double exp10l(long double);
+double exp10l(double);
 
 double      pow10(double);
 float       pow10f(float);
-long double pow10l(long double);
+double pow10l(double);
 #endif
 
 #ifdef __cplusplus
