+++ src/complex/ccoshl.c
@@ -1,7 +1,7 @@
 #include "libm.h"
 
 //FIXME
-long double complex ccoshl(long double complex z)
+double complex ccoshl(double complex z)
 {
 	return ccosh(z);
 }
