+++ src/complex/ccosl.c
@@ -1,12 +1,12 @@
 #include "libm.h"
 
 #if LDBL_MANT_DIG == 53 && LDBL_MAX_EXP == 1024
-long double complex ccosl(long double complex z)
+double complex ccosl(double complex z)
 {
 	return ccos(z);
 }
 #else
-long double complex ccosl(long double complex z)
+double complex ccosl(double complex z)
 {
 	return ccoshl(CMPLXL(-cimagl(z), creall(z)));
 }
