+++ src/complex/cpowl.c
@@ -1,12 +1,12 @@
 #include "libm.h"
 
 #if LDBL_MANT_DIG == 53 && LDBL_MAX_EXP == 1024
-long double complex cpowl(long double complex z, long double complex c)
+double complex cpowl(double complex z, double complex c)
 {
 	return cpow(z, c);
 }
 #else
-long double complex cpowl(long double complex z, long double complex c)
+double complex cpowl(double complex z, double complex c)
 {
 	return cexpl(c * clogl(z));
 }
