+++ src/math/acosl.c
@@ -11,13 +11,13 @@
  */
 /*
  * See comments in acos.c.
- * Converted to long double by David Schultz <das@FreeBSD.ORG>.
+ * Converted to double by David Schultz <das@FreeBSD.ORG>.
  */
 
 #include "libm.h"
 
 #if LDBL_MANT_DIG == 53 && LDBL_MAX_EXP == 1024
-long double acosl(long double x)
+double acosl(double x)
 {
 	return acos(x);
 }
@@ -29,10 +29,10 @@ long double acosl(long double x)
 #define CLEARBOTTOM(u) (u.i.lo = 0)
 #endif
 
-long double acosl(long double x)
+double acosl(double x)
 {
 	union ldshape u = {x};
-	long double z, s, c, f;
+	double z, s, c, f;
 	uint16_t e = u.i.se & 0x7fff;
 
 	/* |x| >= 1 or nan */
