+++ src/math/fdiml.c
@@ -2,12 +2,12 @@
 #include <float.h>
 
 #if LDBL_MANT_DIG == 53 && LDBL_MAX_EXP == 1024
-long double fdiml(long double x, long double y)
+double fdiml(double x, double y)
 {
 	return fdim(x, y);
 }
 #else
-long double fdiml(long double x, long double y)
+double fdiml(double x, double y)
 {
 	if (isnan(x))
 		return x;
